Removed unused modules and components of Tree Borrows.
