+++ bsan-rt/src/borrow_tracker/range_map.rs
@@ -14,7 +14,7 @@ use core::cmp::Ordering;
 use core::ops::Range;
 use core::{iter, mem};
 
-use super::types::Size;
+use super::Size;
 use crate::vec_in;
 
 #[derive(Clone, Debug)]
